Updated the comments

+++ Console_Calculator/BasicCalculator.cs
@@ -6,7 +6,7 @@ namespace Calculator
 	public class BasicCalculator
     {
       //Function to perform basic arithmetic operation
-        public void BasicOperations(double num1, double num2,string op)
+        public void BasicOperations(int num1, int num2,string op)
 		{
             // Use a switch statement to do the math.
             switch (op)
@@ -21,13 +21,14 @@ namespace Calculator
                     Console.WriteLine($"Your result: {num1} * {num2} = " + (num1 * num2));
                     break;
                 case "d":
-                    // Ask the user to enter a non-zero divisor until they do so.
+                    //Ask the user to enter a non - zero divisor until they do so.
                     //while (num2 == 0)
                     //{
                     //    Console.WriteLine("Enter a non-zero divisor: ");
                     //    num2 = Convert.ToInt32(Console.ReadLine());
                     //}
                     Console.WriteLine($"Your result: {num1} / {num2} = " + (num1 / num2));
+
                     break;
             }
              //Wait for the user to respond before closing.
